refactor: Controller와 Model 분리
